requirements.txt
aioredis==1.3.1
APScheduler==3.6.3
asgiref==2.3.0
astroid==2.4.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
channels==2.1.1
channels-redis==2.2.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
daphne==2.1.1
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.2.1
django-crispy-forms==1.7.2
django-friendship==1.8.0
djangorestframework==3.11.0
gitdb==4.0.4
gitdb2==4.0.2
GitPython==3.1.1
hiredis==1.0.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.5.6
Pillow==5.4.1
psycopg2-binary==2.8.5
pylint==2.5.0
PyPDF2==1.26.0
pypiwin32==223
pytz==2018.9
pywin32==227
reportlab==3.5.42
requests==2.23.0
six==1.14.0
smmap==3.0.2
smmap2==2.0.3
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
Twisted==18.4.0
txaio==2.10.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.9
webencodings==0.5.1
wrapt==1.12.1
zope.interface==4.5.0
